// angularapp/src/app/models/loanapplication.model.ts
export interface LoanApplication { 
    loanApplicationId?: number; 
    userId?: number; 
    username?: string; 
    loanId?: number; 
    submissionDate: String; 
    income: number; 
    model: string; 
    purchasePrice: number; 
    loanStatus: number; 
    address: string; 
    file: string; 
}